Update mysql.c: add casts for MSVC 64-bit warnings

+++ source/mysql.c
@@ -171,7 +171,7 @@ HB_FUNC(MYSEXEC)
   assert(session->dbh != SR_NULLPTR);
   session->ulAffected_rows = 0;
   // mysql_query(session->dbh, szQuery);
-  mysql_real_query(session->dbh, szQuery, hb_parclen(2));
+  mysql_real_query(session->dbh, szQuery, (unsigned long)hb_parclen(2));
   session->stmt = mysql_store_result(session->dbh);
   session->ulAffected_rows = mysql_affected_rows(session->dbh);
   if (session->stmt) {
@@ -412,7 +412,7 @@ HB_FUNC(MYSLINEPROCESSED)
     hb_retni(SQL_INVALID_HANDLE);
   } else {
     if (session->ifetch >= -1) {
-      cols = hb_arrayLen(pFields);
+      cols = (int)hb_arrayLen(pFields);
 
       mysql_data_seek(session->stmt, session->ifetch);
       thisrow = mysql_fetch_row(session->stmt);
@@ -730,7 +730,7 @@ HB_FUNC(MYSTABLEATTR)
   sprintf(attcmm, "select * from %s where 0 = 1", hb_parc(2));
 
   // mysql_query(session->dbh, attcmm);
-  mysql_real_query(session->dbh, attcmm, strlen(attcmm));
+  mysql_real_query(session->dbh, attcmm, (unsigned long)strlen(attcmm));
   session->stmt = mysql_store_result(session->dbh);
 
   if (!session->stmt) {
